some tests not passing due to lib updates. ignition scheduling manually tested ok

+++ Control/libs/Reluctor/include/Reluctor/IReluctor.h
@@ -8,7 +8,7 @@ namespace Reluctor
 	{
 	public:
 		virtual float GetPosition() = 0;
-		virtual uint32_t GetTickPerDegree() = 0;
+		virtual float GetTickPerDegree() = 0;
 		virtual uint16_t GetRpm() = 0;
 		virtual uint16_t GetResolution() = 0;
 		virtual bool IsSynced() = 0;
